requirements.txt
guessit
pynzbget